Updated calendar
-updated calendar placeholder
-removed redundant code
